% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotorious.R
\name{annotorious-shiny}
\alias{annotorious-shiny}
\alias{annotoriousOutput}
\alias{renderAnnotorious}
\alias{openseadragonOutput}
\alias{renderOpenSeaDragon}
\title{Shiny bindings for annotorious}
\usage{
annotoriousOutput(outputId, width = "100\%", height = "400px")

renderAnnotorious(expr, env = parent.frame(), quoted = FALSE)

openseadragonOutput(outputId, width = "100\%", height = "400px")

renderOpenSeaDragon(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a annotorious}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
An output element for use in a Shiny user interface.\cr
Consisting of a toggle button to switch between rectangle / polygon mode (id: \code{outputId}\code{-toggle}) and
the html-widget (id: \code{outputId}) which contains an image (id: \code{outputId}\code{-img})
}
\description{
Output and render functions for using annotorious within Shiny
applications and interactive Rmd documents.
}
\examples{
if(interactive() && require(shiny)){
##
## Annotorious
##
library(shiny)
library(recogito)
url <- paste("https://upload.wikimedia.org/",
             "wikipedia/commons/a/a0/Pamphlet_dutch_tulipomania_1637.jpg",
             sep = "")
ui <- fluidPage(annotoriousOutput(outputId = "anno"),
                tags$hr(),
                tags$h3("Results"),
                verbatimTextOutput(outputId = "annotation_result"))
server <- function(input, output) {
  output$anno <- renderAnnotorious({
    annotorious("annotations", tags = c("IMAGE", "TEXT"), src = url)
  })
  output$annotation_result <- renderPrint({
    read_annotorious(input$annotations)
  })
}
shinyApp(ui, server)

##
## Annotorious using OpenSeaDragon, allowing to zoom in,
## to select an area, press shift and next select the area
##
library(shiny)
library(recogito)
url <- paste("https://upload.wikimedia.org/",
             "wikipedia/commons/a/a0/Pamphlet_dutch_tulipomania_1637.jpg",
             sep = "")
ui <- fluidPage(openseadragonOutput(outputId = "anno"),
                tags$hr(),
                tags$h3("Results"),
                verbatimTextOutput(outputId = "annotation_result"))
server <- function(input, output) {
  output$anno <- renderOpenSeaDragon({
    annotorious("annotations", tags = c("IMAGE", "TEXT"), src = url, type = "openseadragon")
  })
  output$annotation_result <- renderPrint({
    read_annotorious(input$annotations)
  })
}
shinyApp(ui, server)

##
## Annotorious using OpenSeaDragon, allowing to zoom in, no selection possibilities
## showing how to load a local image
##
library(shiny)
library(recogito)
url <- system.file(package = "recogito",
                   "examples", "Pamphlet_dutch_tulipomania_1637.jpg")
addResourcePath(prefix = "img", directoryPath = dirname(url))
ui <- fluidPage(openseadragonOutput(outputId = "anno", width = "100\%", height = "250px"))
server <- function(input, output) {
  output$anno <- renderOpenSeaDragon({
    annotorious("annotations", src = sprintf("img/\%s", basename(url)),
                 type = "openseadragon-notoolbar")
  })
}
shinyApp(ui, server)

}
}
