% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\docType{data}
\name{openseadragon_areas}
\alias{openseadragon_areas}
\title{A dataset of annotations using openseadragon}
\value{
\code{x} where column \code{polygon} and the rectangle information in \code{x, y, width, height} is limited to the provided bounding box
}
\description{
A dataset of annotations using openseadragon
}
\examples{
data(openseadragon_areas)
openseadragon_areas
attr(openseadragon_areas, "src")
}
