% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotorious.R
\name{annotorious}
\alias{annotorious}
\title{Annotate images with areas of interest}
\usage{
annotorious(
  inputId = "annotations",
  src,
  tags = c(),
  type = c("annotorious", "openseadragon", "openseadragon-notoolbar"),
  quickselector = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL,
  dependencies = NULL
)
}
\arguments{
\item{inputId}{character string with the name to use where annotations will be put into}

\item{src}{character string with the image/url to annotate}

\item{tags}{character vector of possible labels you want to use}

\item{type}{either \code{'annotorious'}, \code{'openseadragon'}, \code{'openseadragon-notoolbar'} in order to allow zooming with openseadragon or not, with or without a toolbar}

\item{quickselector}{logical indicating if for type \code{'openseadragon'} the possible tags should be shows as quick buttons to click. Defaults to \code{TRUE}.}

\item{width}{passed on to \code{\link[htmlwidgets]{createWidget}}}

\item{height}{passed on to \code{\link[htmlwidgets]{createWidget}}}

\item{elementId}{passed on to \code{\link[htmlwidgets]{createWidget}}}

\item{dependencies}{passed on to \code{\link[htmlwidgets]{createWidget}}}
}
\value{
An object of class htmlwidget as returned by \code{\link[htmlwidgets]{createWidget}}
}
\description{
Functionality to label areas in images
}
\seealso{
\code{\link{annotorious-shiny}}
}
